% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_per_sample}
\alias{data_per_sample}
\title{Metadata per scRNA experiment}
\format{
A data frame with variables: \code{sid}, \code{sample_name},
  \code{doi} and \code{path} to the external data source.
}
\usage{
data_per_sample
}
\description{
Paths etc.
}
\keyword{datasets}
